pyproject.toml
[project]
name = "adk-workflow-framework"
version = "0.1.0"
description = "A framework for orchestrating custom workflows using Google ADK with YAML configuration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.24.0",

]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
workflow = "main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
